fix: use a growable bloom filter implementation.
This commit replaces the fast bloom filter implementation with a growable one.

A growable implementation is required as the bloom filter is used within a
generator; the number of elements in the bloom filter is unbounded and
dependent upon how many docker tasks are spawned rather than how many docker
tasks may spawn concurrently.

Included with these changes:

* Use `thiserror` for custom error types (the crate was already a transitive
  dependency).
* Replace `Option<NonEmpty<C>>` where C is a collection type with the
  collection type itself; this makes the API more ergonomic.
* Replace `args` in `Execution` with `program` and `args`; this makes the API a
  little less confusing as not everything is put into `args`. The name
  `program` was chosen as that's what `std::process::Command::new` uses.
* Deleted a few unused `builder.rs` files.