revise: updates `bollard` to `0.19.0-rc.1`
This commit updates `bollard` to `0.19.0-rc.1`, which is the first commit that
contains @peterhuene's Nodes and Tasks API changes. This also includes some
minor updates to handling of ownership to avoid unnecessary clones (e.g., `impl
AsRef<str>` to `impl Into<String>`).

A good portion of this work was done in #28 and modified to reduce clones.

Closes #22.
Closes #27.
Closes #28.

Co-authored-by: srijanmahajan <srajanmahajan1@gmail.com>