feat: improve backend error handling. (#36)
* feat: improve backend error handling.

This commit enables backends to signal that a task has been cancelled or
preempted to the caller rather than returning a generic error message.

This is specifically needed for detected a TES task was preempted and allow the
caller to handle the preemption by perhaps retrying the task as not
preemptible.

It also updates to the latest dependencies and syncs to the current
implementation of the `tes` crate.

* chore: update CHANGELOG.

* chore: update to `tes` 0.7.0

* chore: code review feedback.

* Use `canceled` consistently.
* Remove `eyre` dependency entirely in favor of `anyhow`.