fix: incorrect resources sizes used by the TES backend. (#37)
* fix: incorrect resources sizes used by the TES backend.

The TES API defines both `ram_gb` and `disk_gb` as having gigabyte (GB) units
while the Crankshaft resources are specified in gibibytes (GiB).

However, the conversion from a Crankshaft resources into a TES resources fails
to convert between the units.

The fix is to convert from GiB to GB for these fields.

This also corrects rounding up `cpu` to its ceiling and removes the limitation
of not specifying zones.

* chore: update CHANGELOG.

* chore: normalize links to PRs in the change logs.