feat: put `crankshaft-monitor` dependency behind a feature gate.
This commit puts the `crankshaft-monitor` dependency of `crankshaft-engine`
behind a `monitoring` feature gate.

To facilitate this, a new crate called `crankshaft-events` was added that
defines a Rust enum for the broadcast event channel. By not directly exposing
the Protobuf types from the Crankshaft API we can now require a Protobuf
compiler at build time only when the `monitoring` feature is enabled.

`crankshaft-engine` was also refactored to pass the event channel down into the
backends directly so that the `spawn` and `run` methods don't need to be
constantly passed the event channel.

These changes contain some tweaks to the `crankshaft-monitor` and
`crankshaft-console` implementations. For the monitor, we now process service
state as its own task independent of a client's request to subscribe to an
events stream. For the console, the "resources" view was removed as retrieving
resources is not supported by all of the Crankshaft backends. The "tasks" view
was tweaked slightly as well.

All dependencies were updated to latest as well.