fix: only subscribe a receiver in the monitor upon request.
This commit changes the implementation of the monitor in `crankshaft-monitor`
to store a `Sender` rather than a `Receiver`. The implementation of broadcast
channels in tokio requires each `Receiver` to read the message from the channel
before removing it from the channel.

By keeping a `Receiver` around that isn't read, the events channel will
needlessly fill to its capacity.

Instead the monitor should be storing the `Sender` and calling `subscribe` upon
a request to subscribe to the events stream.

This also fixes the event streams subscribed by clients to terminate when the
server shuts down.