fix: add user's effective gid for Docker containers.
This commit adds the user's effective gid for containers started by the Docker
backend.

The intention is to allow containers using an image with a `USER` directive to
access the working directory and other mount points under the assumption that
the calling user's group also has access.

This is similar to what `miniwdl` does to allow containers running as a
specific uid to access mounts.