chore: code review feeback and new test.
This commit adds a test for ensuring the Docker backend is adding the caller's
effective group id to the containers it creates.

Replaced the `libc` dependency with `nix`.

Minor refactoring to the builds of `Container` and `Service` to make the name
optional (if one isn't specified, the Docker daemon will use its own randomly
generated name).