fix: fix missing log outputs for Docker backend when in Swarm. (#55)
* fix: fix missing log outputs for Docker backend when in Swarm.

When the local Docker daemon is participating in a swarm, log files were only
being written if the task state becomes "starting" or "running".

However, it is possible for the container to be short lived enough that the
container goes from a pending to a "completed" state.

The completed state wasn't writing the log files.

The fix was to refactor the log writing to share between the container
(non-swarm) and service (swarm) implementations and to call `write_logs` from
seeing a "completed" state.

* chore: update CHANGELOG.

* fix: also fix writing logs for FAILED tasks.

A failed Docker task state occurs for non-zero exit; we should not be treating
these as errors and should instead be writing logs as if the task had completed
with a 0 exit.

Additionally, use `logs` instead of attaching to the container for non-swarm.