fix: add user's effective gid for Docker containers. (#54)
* fix: add user's effective gid for Docker containers.

This commit adds the user's effective gid for containers started by the Docker
backend.

The intention is to allow containers using an image with a `USER` directive to
access the working directory and other mount points under the assumption that
the calling user's group also has access.

This is similar to what `miniwdl` does to allow containers running as a
specific uid to access mounts.

* chore: update CHANGELOGs.

* chore: code review feeback and new test.

This commit adds a test for ensuring the Docker backend is adding the caller's
effective group id to the containers it creates.

Replaced the `libc` dependency with `nix`.

Minor refactoring to the builds of `Container` and `Service` to make the name
optional (if one isn't specified, the Docker daemon will use its own randomly
generated name).