refactor: improve TES backend monitoring of tasks. (#60)
* refactor: improve TES backend monitoring of tasks.

This commit improves the TES backend's monitoring of tasks.

Previously, the TES backend would send a request for each task at the
monitoring interval, which scales linearly with the number of tasks. For a
large number of tasks, it is prohibitively resource intensive to monitor.

These changes instead tag a set of tasks for monitoring and uses the TES
`list_tasks` API to get the status of the tagged tasks with as few network
requests as possible. The tag used to group the tasks changes whenever the
backend is requested to run a task with no tasks currently running.

Only the monitoring tokio task will now periodically wake up for the interval
timer to query the task state; the other tokio tasks wait for a completion
notification from the monitor.

* chore: update CHANGELOG.

* chore: fix doc test.

* Update crankshaft-engine/src/service/runner/backend/tes.rs

Co-authored-by: Adam C. Foltzer <acfoltzer@acfoltzer.net>

* chore: code review feedback.

Refactor the monitoring out into its own module.

* chore: code review feedback.

Ensure `tokio::spawn` is called from async context.

* chore: refactor `select!` in the monitor.

* chore: code review feedback.

---------

Co-authored-by: Adam C. Foltzer <acfoltzer@acfoltzer.net>